going to try running one on the lab machine

+++ reconstruction_runner.py
@@ -2,6 +2,7 @@ import os
 import sys
 
 dry_run = '--dry-run' in sys.argv
+local   = '--local' in sys.argv
 
 if not os.path.exists("slurm_logs"):
     os.makedirs("slurm_logs")
@@ -19,23 +20,23 @@ base_networks = {
 jobs = [
 
         # A couple of quick tests
-        {
-            'datasetdir': 'dataset-trained-more',
-            'num_train_batches': 24000,
-            'num_test_batches': 2800,
-            'epoch_size': 5,
-            'tests_per_epoch': 5,
-            'learning_rate': '-0.00001'
-        },
-        {
-            'datasetdir': 'dataset-trained-more',
-            'num_train_batches': 24000,
-            'num_test_batches': 2800,
-            'grayscale': True,
-            'epoch_size': 5,
-            'tests_per_epoch': 5,
-            'learning_rate': '-0.0001'
-        },
+        # {
+        #     'datasetdir': 'dataset-trained-more',
+        #     'num_train_batches': 24000,
+        #     'num_test_batches': 2800,
+        #     'epoch_size': 5,
+        #     'tests_per_epoch': 5,
+        #     'learning_rate': '-0.00001'
+        # },
+        # {
+        #     'datasetdir': 'dataset-trained-more',
+        #     'num_train_batches': 24000,
+        #     'num_test_batches': 2800,
+        #     'grayscale': True,
+        #     'epoch_size': 5,
+        #     'tests_per_epoch': 5,
+        #     'learning_rate': '-0.0001'
+        # },
 
 
         # the real jobs
@@ -105,7 +106,17 @@ jobs = [
         #     'num_test_batches': 2800,
         #     'grayscale': True,
         #     'learning_rate': '-0.00001'
-        # }
+        # },
+
+        {
+            'datasetdir': 'dataset-trained-more',
+            'num_train_batches': 24000,
+            'num_test_batches': 2800,
+            'dim_hidden': 400,
+            'feature_maps': 128,
+            'grayscale': True,
+            'learning_rate': '-0.0001'
+        },
     ]
 
 if dry_run:
@@ -138,22 +149,21 @@ for job in jobs:
     flagstring = flagstring + " --name " + jobname
 
     jobcommand = "th reconstruction_main.lua" + flagstring
-    with open('slurm_scripts/' + jobname + '.slurm', 'w') as slurmfile:
-        slurmfile.write("#!/bin/bash\n")
-        slurmfile.write("#SBATCH --job-name"+"=" + jobname + "\n")
-        slurmfile.write("#SBATCH --output=slurm_logs/" + jobname + ".out\n")
-        slurmfile.write("#SBATCH --error=slurm_logs/" + jobname + ".err\n")
-        slurmfile.write(jobcommand)
-
-    # if not os.path.exists(jobname):
-    #     os.makedirs(jobname)
-
-    # with open(jobname + '/generating_parameters.txt', 'w') as paramfile:
-    #     paramfile.write(str(job))
 
     print(jobcommand)
-    if not dry_run:
-        os.system("sbatch -N 1 -c 2 --gres=gpu:1 -p gpu --time=6-23:00:00 slurm_scripts/" + jobname + ".slurm &")
+    if local and not dry_run:
+        os.system(jobcommand + ' 2> ' + jobname + '.err 1>' + jobname + '.out')
+
+    else:
+        with open('slurm_scripts/' + jobname + '.slurm', 'w') as slurmfile:
+            slurmfile.write("#!/bin/bash\n")
+            slurmfile.write("#SBATCH --job-name"+"=" + jobname + "\n")
+            slurmfile.write("#SBATCH --output=slurm_logs/" + jobname + ".out\n")
+            slurmfile.write("#SBATCH --error=slurm_logs/" + jobname + ".err\n")
+            slurmfile.write(jobcommand)
+
+        if not dry_run:
+            os.system("sbatch -N 1 -c 2 --gres=gpu:1 -p gpu --time=6-23:00:00 slurm_scripts/" + jobname + ".slurm &")
